Fix ellipsis menu naming
